:recycle: [refactor] (#88) 파일 위치 변경으로 인한 오류 수정
