+++ app/src/main/AndroidManifest.xml
@@ -36,49 +36,51 @@
         android:theme="@style/Theme.MediForMe"
         tools:targetApi="31">
         <activity
-            android:name=".login.SearchmainActivity"
+            android:name=".ui.login.SearchmainActivity"
             android:exported="false" />
         <activity
-            android:name=".login.SearchpwdFragment"
+            android:name=".ui.login.SearchpwdFragment"
+            android:exported="false"
+            tools:ignore="Instantiatable" />
+        <activity
+            android:name=".ui.login.SearchIDFragment"
+            android:exported="false"
+            tools:ignore="Instantiatable" />
+        <activity
+            android:name=".ui.login.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".login.SearchIDFragment"
+            android:name=".ui.join.JoinNameActivity"
             android:exported="false" />
         <activity
-            android:name=".login.LoginActivity"
+            android:name=".ui.join.JoinPasswordActivity"
             android:exported="false" />
         <activity
-            android:name=".join.JoinNameActivity"
+            android:name=".ui.join.JoinServiceActivity"
             android:exported="false" />
         <activity
-            android:name=".join.JoinPasswordActivity"
+            android:name=".ui.join.JoinIdActivity"
             android:exported="false" />
         <activity
-            android:name=".join.JoinServiceActivity"
+            android:name=".ui.join.JoinVericodeActivity"
             android:exported="false" />
         <activity
-            android:name=".join.JoinIdActivity"
+            android:name=".ui.join.JoinPhoneActivity"
             android:exported="false" />
         <activity
-            android:name=".join.JoinVericodeActivity"
-            android:exported="false" />
-        <activity
-            android:name=".join.JoinPhoneActivity"
-            android:exported="false" />
-        <activity
-            android:name=".home.chat.ChatActivity"
+            android:name=".ui.home.chat.ChatActivity"
             android:exported="false" />
         <activity
             android:name=".ui.search.AddMedicineActivity"
             android:exported="false" />
         <activity
-            android:name=".search.AddMedicineResultActivity"
+            android:name=".ui.search.AddMedicineResultActivity"
             android:exported="false" />
         <activity
             android:name=".ui.MainActivity"
             android:exported="false" />
         <activity
-            android:name=".splash.SplashActivity"
+            android:name=".ui.splash.SplashActivity"
             android:windowSoftInputMode="adjustPan"
             android:exported="true">
             <intent-filter>
@@ -88,28 +90,28 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".search.SearchResultActivity"
+            android:name=".ui.search.SearchResultActivity"
             android:exported="false" />
         <activity
-            android:name=".onboarding.OnboardingAgeActivity"
+            android:name=".ui.onboarding.OnboardingAgeActivity"
             android:exported="false" />
         <activity
-            android:name=".search.CameraActivity"
+            android:name=".ui.search.CameraActivity"
             android:exported="false" />
         <activity
-            android:name=".onboarding.OnboardingDetailActivity"
+            android:name=".ui.onboarding.OnboardingDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".search.SearchWithNameActivity"
+            android:name=".ui.search.SearchWithNameActivity"
             android:exported="false" />
         <activity
-            android:name=".onboarding.OnboardingMedicineActivity"
+            android:name=".ui.onboarding.OnboardingMedicineActivity"
             android:exported="true" />
         <activity
-            android:name=".splash.SplashChoiceActivity"
+            android:name=".ui.splash.SplashChoiceActivity"
             android:exported="false" />
         <activity
-            android:name=".search.CheckMedicineActivity"
+            android:name=".ui.search.CheckMedicineActivity"
             android:exported="false" />
     </application>
