+++ app/src/main/res/layout/activity_join_id.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".join.JoinIdActivity">
+    tools:context=".ui.join.JoinIdActivity">
 
     <ImageView
         android:id="@+id/main_logo_IV"
