+++ app/src/main/res/layout/activity_join_name.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".join.JoinNameActivity">
+    tools:context=".ui.join.JoinNameActivity">
 
     <ImageView
         android:id="@+id/main_logo_IV"
