+++ app/src/main/res/layout/activity_join_password.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".join.JoinPasswordActivity">
+    tools:context=".ui.join.JoinPasswordActivity">
 
     <ImageView
         android:id="@+id/main_logo_IV"
