+++ app/src/main/res/layout/activity_join_service.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".join.JoinServiceActivity">
+    tools:context=".ui.join.JoinServiceActivity">
 
     <ImageView
         android:id="@+id/main_logo_IV"
