+++ app/src/main/res/layout/activity_searchmain.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".login.SearchmainActivity">
+    tools:context=".ui.login.SearchmainActivity">
 
     <ImageView
         android:id="@+id/main_logo_IV"
