+++ app/src/main/res/layout/activity_splash_init.xml
@@ -5,8 +5,8 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/back"
-    tools:context=".SplashInit">
+    android:background="@color/back">
+<!--    tools:context=".SplashInit">-->
 
     <ImageView
         android:id="@+id/logo_img_IV"
