+++ app/src/main/res/layout/fragment_infoid.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.InfoIDFragment">
+    tools:context=".ui.login.InfoIDFragment">
 
     <TextView
         android:id="@+id/fst_TV"
