+++ app/src/main/res/layout/fragment_infopwd.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".login.InfoPWDFragment">
+    tools:context=".ui.login.InfoPWDFragment">
 
     <TextView
         android:id="@+id/fst_TV"
