+++ app/src/main/res/layout/fragment_searchid.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".login.SearchIDFragment">
+    tools:context=".ui.login.SearchIDFragment">
 
     <EditText
         android:id="@+id/phone_num_ET"
