+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.example.mediforme.home.HomeFragment"
+        android:name="com.example.mediforme.ui.home.HomeFragment"
         android:label="HomeFragment"
         tools:layout="@layout/fragment_home"
         >
@@ -21,27 +21,27 @@
 
     <fragment
         android:id="@+id/todayConditionFragment"
-        android:name="com.example.mediforme.home.todayCondition.TodayConditionFragment"
+        android:name="com.example.mediforme.ui.home.todayCondition.TodayConditionFragment"
         android:label="fragment_today_condition"
         tools:layout="@layout/fragment_today_condition">
     </fragment>
 
     <fragment
         android:id="@+id/chatFragment"
-        android:name="com.example.mediforme.home.chat.ChatFragment"
+        android:name="com.example.mediforme.ui.home.chat.ChatFragment"
         android:label="ChatFragment"
         tools:layout="@layout/fragment_chat"/>
 
     <fragment
         android:id="@+id/navigation_search"
-        android:name="com.example.mediforme.search.SearchFragment"
+        android:name="com.example.mediforme.ui.search.SearchFragment"
         android:label="SearchFragment"
         tools:layout="@layout/fragment_search"/>
 
 
     <fragment
         android:id="@+id/navigation_mypage"
-        android:name="com.example.mediforme.mypage.MyPageFragment"
+        android:name="com.example.mediforme.ui.mypage.MyPageFragment"
         android:label="MyPageFragment"
         tools:layout="@layout/fragment_mypage"
         />
